feat: Rewrite components to be dockerized
BREAKING CHANGE: The kapellmeister is now going obsolete. PA
components are now to be started manually or with a separate script.

+++ router.py
@@ -7,7 +7,8 @@ import signal
 import socket
 import sys
 
-from core import Config, Kapellmeister
+from channels import SocketChannel
+
 from core.routing import Router
 
 
@@ -15,19 +16,28 @@ def _term(*_):
     exit()
 
 
-def main(sockname, config_file_name):
-    with open(config_file_name) as cfg:
-        config = Config(cfg.read())
-    km = Kapellmeister(config)
-    km.run()
-    be = km.connect("brain")
+def start_server(sockname):
+    if ':' in sockname:
+        host, port = sockname.split(':')
+        serv = socket.socket(socket.AF_INET)
+        serv.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEPORT, 1)
+        serv.bind((host, int(port)))
+        sockname = serv.getsockname()
+        sys.stdout.flush()
+    else:
+        serv = socket.socket(socket.AF_UNIX)
+        serv.bind(sockname)
+        atexit.register(os.unlink, sockname)
+    serv.listen(0)
+    return serv, sockname
 
-    sock = socket.socket(socket.AF_UNIX)
-    sock.bind(sockname)
-    atexit.register(os.unlink, sockname)
 
-    sock.listen()
+def main(sockname, config_file_name, brain_sockname):
+    brain_serv, brain_serv_addr = start_server(brain_sockname)
 
+    be = SocketChannel(brain_serv.accept()[0])
+
+    sock, router_addr = start_server(sockname)
     router = Router(sock, be)
 
     while True:
@@ -40,6 +50,7 @@ if __name__ == '__main__':
 
     parser = argparse.ArgumentParser()
     parser.add_argument('-s', '--socket', required=True)
+    parser.add_argument('-b', '--brain-socket', default="0.0.0.0:0")
     parser.add_argument('-c', '--config', default="config.yml")
     args = parser.parse_args()
-    main(args.socket, args.config)
+    main(args.socket, args.config, args.brain_socket)
